requirements.txt
aiofiles==23.1.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
fastapi==0.104.0
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
magic-filter==1.0.12
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
yarl==1.9.2
